// frontend/src/types/generated/models/PromptPreviewResponse.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { PromptPreviewResult } from "./PromptPreviewResult";

/**
 * 预览执行响应
 */
export type PromptPreviewResponse = { results: Array<PromptPreviewResult>, totalPassed: number, totalFailed: number, totalExecutionTimeMs: number, };